backend/requirements.txt
fastapi
uvicorn
llama-cpp-python
diskcache
psutil
pydantic
huggingface_hub
python-multipart
